node_port_val added to function
