+++ ansible_collections/arista/avd/plugins/plugin_utils/topology_generator_utils.py
@@ -10,6 +10,13 @@ import textwrap
 
 import drawSvg as draw
 
+import re
+
+_nsre = re.compile('([0-9]+)')
+def natural_sort_key(s):
+    return [int(text) if text.isdigit() else text.lower()
+            for text in re.split(_nsre, s)] 
+
 # MARGINS between real groups
 BOXMARGIN = 25
 
@@ -713,3 +720,77 @@ def draw_same_level_links(d, node, neigh, nol, nel, stroke, used_height):
     d.append(draw.Line(nel["x"], nel["y"], nel["x"] - XOffset, nel["y"], stroke=stroke, stroke_width=2))
     d.append(draw.Line(nel["x"] - XOffset, nel["y"], nel["x"] - XOffset, routey, stroke=stroke, stroke_width=2))
     d.append(draw.Line(nel["x"] - XOffset, routey, routex, routey, stroke=stroke, stroke_width=2))
+
+
+def get_node_port_temp_graph_dict(graph_dict, node_level_dict, level_dict, undefined_rank_nodes):
+    node_port_val = {}
+
+    for i in node_level_dict.keys():
+        if i not in undefined_rank_nodes:
+            node_port_val[i] = {}
+            node_port_val[i]["checked"] = []
+            node_port_val[i]["top"] = []
+            node_port_val[i]["bottom"] = []
+            node_port_val[i]["left"] = []
+            node_port_val[i]["right"] = []
+
+    # avoid same node neighbor pair
+    check_same_node = []
+    temp_graph_dict = {}
+
+    for node_val, node_details in graph_dict.items():
+        if node_val not in undefined_rank_nodes:
+
+            for i in node_details:
+                node_neighbor_str = None
+
+                node_neighbor_str = [node_val] + [i["nodePort"]] + [i["neighborDevice"]] + [i["neighborPort"]]
+
+                node_neighbor_str.sort()
+                node_neighbor_str = "_".join(node_neighbor_str)
+
+                if node_neighbor_str not in check_same_node:
+                    check_same_node.append(node_neighbor_str)
+                    if node_val not in temp_graph_dict.keys():
+                        temp_graph_dict[node_val] = [i]
+                    else:
+                        temp_graph_dict[node_val] = temp_graph_dict[node_val] + [i]
+
+                #node_level_dict
+                # set top and bottom port values
+                if node_level_dict[node_val] < node_level_dict[i["neighborDevice"]]:
+                    if i["nodePort"] not in node_port_val[node_val]["bottom"] and i["nodePort"] and i["nodePort"] not in node_port_val[node_val]["checked"]:
+                        node_port_val[node_val]["bottom"] = node_port_val[node_val]["bottom"] + [i["nodePort"]]
+                        node_port_val[node_val]["checked"] = node_port_val[node_val]["checked"] + [i["nodePort"]]
+
+                    if i["neighborPort"] not in node_port_val[i["neighborDevice"]]["top"] and i["neighborPort"] and i["neighborPort"] not in node_port_val[i["neighborDevice"]]["checked"]:
+                        node_port_val[i["neighborDevice"]]["top"] = node_port_val[i["neighborDevice"]]["top"] + [i["neighborPort"]]
+                        node_port_val[i["neighborDevice"]]["checked"] = node_port_val[i["neighborDevice"]]["checked"] + [i["neighborPort"]]
+
+                if node_level_dict[node_val] > node_level_dict[i["neighborDevice"]]:
+                    if i["nodePort"] not in node_port_val[node_val]["top"] and i["nodePort"] and i["nodePort"] not in node_port_val[node_val]["checked"]:
+                        node_port_val[node_val]["top"] = node_port_val[node_val]["top"] + [i["nodePort"]]
+                        node_port_val[node_val]["checked"] = node_port_val[node_val]["checked"] + [i["nodePort"]]
+
+                    if i["neighborPort"] not in node_port_val[i["neighborDevice"]]["bottom"] and i["neighborPort"] and i["neighborPort"] not in node_port_val[i["neighborDevice"]]["checked"]:
+                        node_port_val[i["neighborDevice"]]["bottom"] = node_port_val[i["neighborDevice"]]["bottom"] + [i["neighborPort"]]
+                        node_port_val[i["neighborDevice"]]["checked"] = node_port_val[i["neighborDevice"]]["checked"] + [i["neighborPort"]]
+
+    for _,level_list in level_dict.items():
+        level_list.sort(key=natural_sort_key)
+
+        for i in range(len(level_list) - 1):
+            for j in range(i+1,len(level_list)): 
+                for node_detail in graph_dict[level_list[i]]:
+                    if node_detail["neighborDevice"] == level_list[j]:
+                        #left node => right port    
+                        if (node_detail["nodePort"] not in node_port_val[level_list[i]]["right"]) and (node_detail["nodePort"] not in node_port_val[level_list[i]]["checked"]):
+                            node_port_val[level_list[i]]["right"] = node_port_val[level_list[i]]["right"] + [node_detail["nodePort"]]
+                            node_port_val[level_list[i]]["checked"] = node_port_val[level_list[i]]["checked"] + [node_detail["nodePort"]]
+                        
+                        #right node => left port 
+                        if (node_detail["neighborPort"] not in node_port_val[level_list[j]]["left"]) and (node_detail["neighborPort"] not in node_port_val[level_list[j]]["checked"]):
+                            node_port_val[level_list[j]]["left"] = node_port_val[level_list[j]]["left"] + [node_detail["neighborPort"]]
+                            node_port_val[level_list[j]]["checked"] = node_port_val[level_list[j]]["checked"] + [node_detail["neighborPort"]]
+
+    return node_port_val, temp_graph_dict
