+++ ansible_collections/arista/avd/roles/topology_generator/tasks/main.yml
@@ -7,7 +7,7 @@
     - name: Generate topology diagram from structure config in svg format
       tags: [always]
       arista.avd.topology_generator:
-        structured_config: "./tests/structured_configs_l2ls_fabric"
+        structured_config: "./tests/structured_configs_twodc_5stage_clos"
         destination: "./example.svg"
       delegate_to: localhost
